Refine guard link handling and tests
