+++ detrackify_guard.py
@@ -10,7 +10,7 @@
 #
 # You should have received a copy of the GNU General Public License along with this program.
 # If not, see <https://www.gnu.org/licenses/>.
-"""Simple Detrackify guard server."""
+"""Detrackify guard server."""
 
 import argparse
 import base64
