+++ templates/guard_warning.html
@@ -27,10 +27,12 @@ setTimeout(enable, {{ timeout_ms }});
 {% else %}
 <p>Invalid link information provided.</p>
 {% endif %}
+{% if valid %}
 <form method="post" action="" autocomplete="off">
   <input type="hidden" name="ts" value="{{ ts }}">
   <button id="cont" disabled>Yes, I understand and want to continue</button>
 </form>
+{% endif %}
 </body>
 </html>
