+++ templates/guard_warning_de.html
@@ -26,10 +26,12 @@ setTimeout(enable, {{ timeout_ms }});
 {% else %}
 <p>Ungültiger Link angegeben.</p>
 {% endif %}
+{% if valid %}
 <form method="post" action="" autocomplete="off">
   <input type="hidden" name="ts" value="{{ ts }}">
   <button id="cont" disabled>Ja, ich verstehe und möchte fortfahren</button>
 </form>
+{% endif %}
 </body>
 </html>
