+++ templates/guard_warning_fr.html
@@ -26,10 +26,12 @@ setTimeout(enable, {{ timeout_ms }});
 {% else %}
 <p>Le lien fourni est invalide.</p>
 {% endif %}
+{% if valid %}
 <form method="post" action="" autocomplete="off">
   <input type="hidden" name="ts" value="{{ ts }}">
   <button id="cont" disabled>Oui, je comprends et je veux continuer</button>
 </form>
+{% endif %}
 </body>
 </html>
