+++ templates/guard_warning_zh.html
@@ -26,10 +26,12 @@ setTimeout(enable, {{ timeout_ms }});
 {% else %}
 <p>提供的链接无效。</p>
 {% endif %}
+{% if valid %}
 <form method="post" action="" autocomplete="off">
   <input type="hidden" name="ts" value="{{ ts }}">
   <button id="cont" disabled>是的，我理解并继续</button>
 </form>
+{% endif %}
 </body>
 </html>
